requirements.txt
datasets
pandas
numpy
scipy
scikit-learn
mlflow
autogluon.tabular
flask
requests
openpyxl

